rewrite movedown()
I abandon the idea of moving down within a line.
Moving down now takes you to the actuall next string, not next terminal line.
The behaviour is now identical to vi.

As far as I can tell, this doesn't introduce any bugs.
But as this is somewhat complex, it's difficult to know.